requirements.txt
httpx
fastapi
uvicorn
websockets
python-dotenv